refactor: share MersenneTwister, use uuid crate for formatting
Extract SparkMersenneTwister into internal/mersenne.rs so shuffle and uuid both
depend on it rather than uuid reaching into the shuffle module. Format UUIDs via
the uuid crate encoded into a pre-sized StringBuilder, removing per-row String
allocation and the hand-rolled hex formatting.
